+++ src/main/java/bot/repository/AllianceMemberRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import bot.entity.AllianceMember;
 
 @Repository
-public interface  AllianceMemberRepository extends JpaRepository<AllianceMember, Long>{
+public interface  AllianceMemberRepository extends JpaRepository<AllianceMember, Integer>{
 	public AllianceMember findByAyarabuName(String ayarabuName);
 	public AllianceMember findByDiscordName(String discordName);
 	public AllianceMember findByDiscordMemberId(String discordMemberId);
