+++ src/main/java/bot/repository/LevelMasterRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import bot.entity.LevelMaster;
 
-public interface LevelMasterRepository extends JpaRepository<LevelMaster, Long>{
+public interface LevelMasterRepository extends JpaRepository<LevelMaster, Integer>{
 
 }
